requirements_fixed.txt
asgiref==3.8.1
attrs==24.3.0
Authlib==1.6.9
Automat==24.8.1
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.3.1
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.0
cssselect==1.2.0
cycler==0.12.1
db-dtypes==1.5.1
defusedxml==0.7.1
Django==5.1.4
dotenv==0.9.9
et_xmlfile==2.0.0
filelock==3.16.1
Flask==3.1.3
fonttools==4.55.3
google-api-core==2.30.3
google-auth==2.49.1
google-auth-oauthlib==1.3.0
google-cloud-bigquery==3.41.0
google-cloud-core==2.5.1
google-crc32c==1.8.0
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
greenlet==3.1.1
grpcio==1.80.0
grpcio-status==1.80.0
gspread==6.2.1
h11==0.14.0
httplib2==0.31.2
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.10.0
itemloaders==1.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
kiwisolver==1.4.8
lxml==5.3.0
MarkupSafe==3.0.3
matplotlib==3.10.0
numpy==2.2.1
oauth2client==4.1.3
oauthlib==3.3.1
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
parsel==1.9.1
pillow==11.1.0
Protego==0.3.1
proto-plus==1.27.2
protobuf==6.33.6
pyarrow==23.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyDispatcher==2.0.7
pyodbc==5.2.0
pyOpenSSL==24.3.0
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2024.2
queuelib==1.7.0
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rsa==4.9.1
Scrapy==2.12.0
selenium==4.27.1
service-identity==24.2.0
setuptools==75.7.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.3
tldextract==5.1.3
trio==0.28.0
trio-websocket==0.11.1
Twisted==24.11.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uv==0.11.3
w3lib==2.2.1
websocket-client==1.8.0
Werkzeug==3.1.7
wsproto==1.2.0
zope.interface==7.2
